Create basicEntity
This class is meant for the creation of simple entities in simple projects. If you intend to build a sophisticated software you shouldn’t use this class as the users data is stored as one type (collection) without data individual access modification. But for smaller projects or MVPs you save time.